move RFT/RPT/RST keyword handlers to separate file
