+++ CMakeLists_files.cmake
@@ -203,6 +203,7 @@ if(ENABLE_ECL_INPUT)
     src/opm/input/eclipse/Schedule/RFTConfig.cpp
     src/opm/input/eclipse/Schedule/RPTConfig.cpp
     src/opm/input/eclipse/Schedule/RSTConfig.cpp
+    src/opm/input/eclipse/Schedule/RXXKeywordHandlers.cpp
     src/opm/input/eclipse/Schedule/Schedule.cpp
     src/opm/input/eclipse/Schedule/ScheduleBlock.cpp
     src/opm/input/eclipse/Schedule/ScheduleDeck.cpp
