+++ src/opm/input/eclipse/Schedule/RXXKeywordHandlers.hpp
@@ -1,0 +1,35 @@
+/*
+  Copyright 2020 Statoil ASA.
+
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef RXX_KEYWORD_HANDLERS_HPP
+#define RXX_KEYWORD_HANDLERS_HPP
+
+#include "KeywordHandlers.hpp"
+
+#include <string>
+#include <utility>
+#include <vector>
+
+namespace Opm {
+
+//! \brief Obtain a list of RFT/RPT/RST keyword handlers.
+std::vector<std::pair<std::string,KeywordHandlers::handler_function>> getRXXHandlers();
+
+}
+
+#endif
